TRIVIAL: Change error msg to warn for Fetchers
Summary:
This moves the log line from an error to a warning. Each implementation
of this thread can logs out when an error or fatal error occurs.

ReplicaFetcherThread, has both warn, error and fatal in the
handleOffsetOutOfRange method.

ConsumerFetcherThread seems to reset itself without logging out an error.

Seems that the Reset message isn't shouldn't be an error level as it
doesn't indicate any real error.

This Error message has occurred during reassignment of partitions and
doesn't indicate anything that is really wrong.